+++ tests/Integrations/Guzzle/V5/GuzzleIntegrationTest.php
@@ -215,4 +215,22 @@ final class GuzzleIntegrationTest extends IntegrationTestCase
         // existing headers are honored
         $this->assertSame('preserved_value', $found['headers']['Honored']);
     }
+
+    public function testAppendHostnameToServiceName()
+    {
+        putenv('DD_TRACE_HTTP_CLIENT_SPLIT_BY_DOMAIN=true');
+
+        $traces = $this->isolateTracer(function () {
+            $this->getMockedClient()->get('http://example.com');
+        });
+        $this->assertSpans($traces, [
+            SpanAssertion::build('GuzzleHttp\Client.send', 'example.com', 'http', 'send')
+                ->setTraceAnalyticsCandidate()
+                ->withExactTags([
+                    'http.method' => 'GET',
+                    'http.url' => 'http://example.com',
+                    'http.status_code' => '200',
+                ]),
+        ]);
+    }
 }